rename layouts folder to assets to remove layouts warning message
