+++ assets/style.scss
@@ -5,7 +5,7 @@
 @import url("https://fonts.googleapis.com/icon?family=Material+Icons");
 
 @import url('reveal.js/dist/reveal.css');
-@import url('~/layouts/nlesc-decorations.scss');
+@import url('~/assets/nlesc-decorations.scss');
 
 p, ul, ol{
   a{
